app/datapipeline/requirements.txt
anthropic[bedrock]>=0.25.0
msal>=1.28.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0

# dev / test
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
